+++ tests/integration/test_attachable.py
@@ -1,31 +1,16 @@
 import os
 import tempfile
-import unittest
 from datetime import datetime
 
-from quickbooks.auth import Oauth1SessionManager
-from quickbooks.client import QuickBooks
 from quickbooks.objects.attachable import Attachable
-from quickbooks.objects.base import Ref, AttachableRef
+from quickbooks.objects.base import AttachableRef
 from quickbooks.objects.vendor import Vendor
+from tests.integration.test_base import QuickbooksTestCase
 
 
-class AttachableTest(unittest.TestCase):
+class AttachableTest(QuickbooksTestCase):
     def setUp(self):
-        self.session_manager = Oauth1SessionManager(
-            sandbox=True,
-            consumer_key=os.environ.get('CONSUMER_KEY'),
-            consumer_secret=os.environ.get('CONSUMER_SECRET'),
-            access_token=os.environ.get('ACCESS_TOKEN'),
-            access_token_secret=os.environ.get('ACCESS_TOKEN_SECRET'),
-        )
-
-        self.qb_client = QuickBooks(
-            session_manager=self.session_manager,
-            sandbox=True,
-            company_id=os.environ.get('COMPANY_ID')
-        )
-
+        super(AttachableTest, self).setUp()
         self.time = datetime.now()
 
     def test_create_note(self):