+++ tests/integration/test_customer.py
@@ -1,30 +1,13 @@
 from datetime import datetime
-import os
-import unittest
 
-from quickbooks.auth import Oauth1SessionManager
 from quickbooks.objects.base import Address, PhoneNumber, EmailAddress
-
 from quickbooks.objects.customer import Customer
+from tests.integration.test_base import QuickbooksTestCase
 
-from quickbooks import QuickBooks
 
-
-class CustomerTest(unittest.TestCase):
+class CustomerTest(QuickbooksTestCase):
     def setUp(self):
-        self.session_manager = Oauth1SessionManager(
-            sandbox=True,
-            consumer_key=os.environ.get('CONSUMER_KEY'),
-            consumer_secret=os.environ.get('CONSUMER_SECRET'),
-            access_token=os.environ.get('ACCESS_TOKEN'),
-            access_token_secret=os.environ.get('ACCESS_TOKEN_SECRET'),
-        )
-
-        self.qb_client = QuickBooks(
-            session_manager=self.session_manager,
-            sandbox=True,
-            company_id=os.environ.get('COMPANY_ID')
-        )
+        super(CustomerTest, self).setUp()
 
         self.title = "Mr"
         self.given_name = "James"