+++ tests/integration/test_purchaseorder.py
@@ -1,8 +1,3 @@
-import os
-import unittest
-
-from quickbooks.auth import Oauth1SessionManager
-from quickbooks.client import QuickBooks
 from quickbooks.objects.account import Account
 from quickbooks.objects.customer import Customer
 from quickbooks.objects.detailline import ItemBasedExpenseLine, ItemBasedExpenseLineDetail
@@ -10,24 +5,10 @@ from quickbooks.objects.item import Item
 from quickbooks.objects.purchaseorder import PurchaseOrder
 from quickbooks.objects.taxcode import TaxCode
 from quickbooks.objects.vendor import Vendor
+from tests.integration.test_base import QuickbooksTestCase
 
 
-class PurchaseOrderTest(unittest.TestCase):
-    def setUp(self):
-        self.session_manager = Oauth1SessionManager(
-            sandbox=True,
-            consumer_key=os.environ.get('CONSUMER_KEY'),
-            consumer_secret=os.environ.get('CONSUMER_SECRET'),
-            access_token=os.environ.get('ACCESS_TOKEN'),
-            access_token_secret=os.environ.get('ACCESS_TOKEN_SECRET'),
-        )
-
-        self.qb_client = QuickBooks(
-            session_manager=self.session_manager,
-            sandbox=True,
-            company_id=os.environ.get('COMPANY_ID')
-        )
-
+class PurchaseOrderTest(QuickbooksTestCase):
     def test_create(self):
         customer = Customer.all(max_results=1, qb=self.qb_client)[0]
         taxcode = TaxCode.all(max_results=1, qb=self.qb_client)[0]