+++ CHANGELOG.rst
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+* 0.8.1 (September 18th, 2019)
+    * Dropped support for Python 2.7 and 3.3
+    * Updated the Invoice to include an AllowOnlineCreditCardPayment attribute.
+    * Updated the SendMixin class to url encode input emails.
+
+* 0.8 (June 25th, 2019)
+    * Removed support for OAuth 1.0
+    * Replace OAuth Session Manager and CLI with intuit-oauth client.
+    * Removed disconnect_account and reconnect_account from client.
+    * Fixed on Invoice object that caused the DocNumber to be set to an empty string.
+    * Added to_ref method to PaymentMethod object.
+    * Added CompanyCurrency object.
+    * Fixed issue that prevented creation of TaxAgencies.
+    * Fixed issues with GroupLine, SubtotalLine, and DescriptionOnlyLine objects.
+    * Fixed issue with CDC when there are no changes within a given timestamp.
+
 * 0.7.5 (October 18th, 2018)
     * Fixed bug with reporting authentication failure when attempting to download PDF (previously the error details were "lost").
     * Added refresh_access_tokens to Oauth2SessionManager.